+++ tests/xharness/AppRunner.cs
@@ -8,15 +8,14 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Xml;
-using Xamarin;
-using Xharness.Hardware;
 using Xharness.Execution;
 using Xharness.Jenkins.TestTasks;
 using Xharness.Listeners;
 using Xharness.Logging;
 using Xharness.Utilities;
 
-namespace Xharness {
+namespace Xharness
+{
 	public enum AppRunnerTarget
 	{
 		None,
@@ -38,15 +37,6 @@ namespace Xharness {
 
 	class AppRunner
 	{
-		string appPath;
-		string launchAppPath;
-		string platform;
-		Extension? extension;
-		bool isSimulator;
-		string configuration;
-		string mode;
-		bool initialized;
-
 		public Harness Harness;
 		public string ProjectFile;
 		public string AppPath;
@@ -55,75 +45,111 @@ namespace Xharness {
 		public ISimpleListenerFactory ListenerFactory = new SimpleListenerFactory ();
 
 		public TestExecutingResult Result { get; private set; }
-
 		public string FailureMessage { get; private set; }
 
-		public string DeviceName { get; set; }
+		string appName;
+		string appPath;
+		string launchAppPath;
+		string bundle_identifier;
+		string platform;
+		Extension? extension;
+		bool isSimulator;
 
-		public string CompanionDeviceName { get; set; }
+		string device_name;
+		string companion_device_name;
 
-		public bool IsExtension => extension.HasValue;
+		string configuration;
+		public string Configuration {
+			get { return configuration ?? Harness.Configuration; }
+			set { configuration = value; }
+		}
 
-		public string AppName { get; private set; }
+		public string DeviceName {
+			get { return device_name; }
+			set { device_name = value; }
+		}
+
+		public string CompanionDeviceName {
+			get { return companion_device_name; }
+			set { companion_device_name = value; }
+		}
+
+		public bool isExtension {
+			get {
+				return extension.HasValue;
+			}
+		}
+
+		public string AppName => appName;
 
 		public double TimeoutMultiplier { get; set; } = 1;
 
-		ILogs logs;
-		public ILogs Logs => logs ?? (logs = new Logs (LogDirectory));
-
-		public ILog MainLog { get; set; }
-
-		public ISimulatorDevice [] Simulators { get; set; }
-		ISimulatorDevice simulator => Simulators [0];
-
-		public string BundleIdentifier { get; private set; }
-
-		public IProcessManager ProcessManager { get; set; } = new ProcessManager ();
+		// For watch apps we end up with 2 simulators, the watch simulator (the main one), and the iphone simulator (the companion one).
+		SimDevice[] simulators;
+		SimDevice simulator { get { return simulators [0]; } }
+		SimDevice companion_simulator { get { return simulators.Length == 2 ? simulators [1] : null; } }
 
 		AppRunnerTarget target;
 		public AppRunnerTarget Target {
-			get => target == AppRunnerTarget.None ? Harness.Target : target;
-			set => target = value;
+			get { return target == AppRunnerTarget.None ? Harness.Target : target; }
+			set { target = value; }
 		}
 
 		string log_directory;
 		public string LogDirectory {
-			get => log_directory ?? Harness.LogDirectory;
-			set => log_directory = value;
+			get { return log_directory ?? Harness.LogDirectory; }
+			set { log_directory = value; }
 		}
 
-		bool ensure_clean_simulator_state = true;
-		public bool EnsureCleanSimulatorState {
-			get => ensure_clean_simulator_state && string.IsNullOrEmpty (Environment.GetEnvironmentVariable ("SKIP_SIMULATOR_SETUP"));
-			set => ensure_clean_simulator_state = value;
+		ILogs logs;
+		public ILogs Logs {
+			get {
+				return logs ?? (logs = new Logs (LogDirectory));
+			}
 		}
 
-		public string Configuration {
-			get => configuration ?? Harness.Configuration;
-			set => configuration = value;
+		ILog main_log;
+		public ILog MainLog {
+			get { return main_log; }
+			set { main_log = value; }
 		}
+
+		public SimDevice [] Simulators {
+			get { return simulators; }
+			set { simulators = value; }
+		}
+
+		public string BundleIdentifier {
+			get {
+				return bundle_identifier;
+			}
+		}
+
+		public IProcessManager ProcessManager { get; set; } = new ProcessManager ();
+
+		string mode;
 
 		async Task<bool> FindSimulatorAsync ()
 		{
-			if (Simulators != null)
+			if (simulators != null)
 				return true;
 			
 			var sims = new Simulators () {
 				Harness = Harness,
 			};
 			await sims.LoadAsync (Logs.Create ($"simulator-list-{Harness.Timestamp}.log", "Simulator list"));
-			Simulators = await sims.FindAsync (Target, MainLog);
+			simulators = await sims.FindAsync (Target, main_log);
 
-			return Simulators != null;
+			return simulators != null;
 		}
 
 		void FindDevice ()
 		{
-			if (DeviceName != null)
+			if (device_name != null)
 				return;
 			
-			DeviceName = Environment.GetEnvironmentVariable ("DEVICE_NAME");
-			if (!string.IsNullOrEmpty (DeviceName))
+			device_name = Environment.GetEnvironmentVariable ("DEVICE_NAME");
+			if (!string.IsNullOrEmpty (device_name))
 				return;
 
 			var devs = new Devices () {
@@ -131,7 +157,7 @@ namespace Xharness {
 			};
 			Task.Run (async () =>
 			{
-				await devs.LoadAsync (MainLog);
+				await devs.LoadAsync (main_log);
 			}).Wait ();
 
 			string [] deviceClasses;
@@ -150,28 +176,30 @@ namespace Xharness {
 			}
 
 			var selected = devs.ConnectedDevices.Where ((v) => deviceClasses.Contains (v.DeviceClass) && v.IsUsableForDebugging != false);
-			IHardwareDevice selected_data;
+			Device selected_data;
 			if (selected.Count () == 0) {
 				throw new Exception ($"Could not find any applicable devices with device class(es): {string.Join (", ", deviceClasses)}");
 			} else if (selected.Count () > 1) {
 				selected_data = selected
 					.OrderBy ((dev) =>
 					{
-						if (Version.TryParse (dev.ProductVersion, out Version v))
+						Version v;
+						if (Version.TryParse (dev.ProductVersion, out v))
 							return v;
 						return new Version ();
 					})
 					.First ();
-				MainLog.WriteLine ("Found {0} devices for device class(es) '{1}': '{2}'. Selected: '{3}' (because it has the lowest version).", selected.Count (), string.Join ("', '", deviceClasses), string.Join ("', '", selected.Select ((v) => v.Name).ToArray ()), selected_data.Name);
+				main_log.WriteLine ("Found {0} devices for device class(es) '{1}': '{2}'. Selected: '{3}' (because it has the lowest version).", selected.Count (), string.Join ("', '", deviceClasses), string.Join ("', '", selected.Select ((v) => v.Name).ToArray ()), selected_data.Name);
 			} else {
 				selected_data = selected.First ();
 			}
-			DeviceName = selected_data.Name;
+			device_name = selected_data.Name;
 
 			if (mode == "watchos")
-				CompanionDeviceName = devs.FindCompanionDevice (MainLog, selected_data).Name;
+				companion_device_name = devs.FindCompanionDevice (main_log, selected_data).Name;
 		}
 
+		bool initialized;
 		public void Initialize ()
 		{
 			if (initialized)
@@ -180,11 +208,11 @@ namespace Xharness {
 
 			var csproj = new XmlDocument ();
 			csproj.LoadWithoutNetworkAccess (ProjectFile);
-			AppName = csproj.GetAssemblyName ();
+			appName = csproj.GetAssemblyName ();
 			var info_plist_path = csproj.GetInfoPListInclude ();
 			var info_plist = new XmlDocument ();
 			info_plist.LoadWithoutNetworkAccess (Path.Combine (Path.GetDirectoryName (ProjectFile), info_plist_path.Replace ('\\', '/')));
-			BundleIdentifier = info_plist.GetCFBundleIdentifier ();
+			bundle_identifier = info_plist.GetCFBundleIdentifier ();
 
 			var extensionPointIdentifier = info_plist.GetNSExtensionPointIdentifier ();
 			if (!string.IsNullOrEmpty (extensionPointIdentifier))
@@ -235,7 +263,7 @@ namespace Xharness {
 				throw new Exception (string.Format ("Unknown target: {0}", Harness.Target));
 			}
 
-			appPath = Path.Combine (Path.GetDirectoryName (ProjectFile), csproj.GetOutputPath (platform, Configuration).Replace ('\\', '/'), AppName + (IsExtension ? ".appex" : ".app"));
+			appPath = Path.Combine (Path.GetDirectoryName (ProjectFile), csproj.GetOutputPath (platform, Configuration).Replace ('\\', '/'), appName + (isExtension ? ".appex" : ".app"));
 			if (!Directory.Exists (appPath))
 				throw new Exception (string.Format ("The app directory {0} does not exist. This is probably a bug in the test harness.", appPath));
 
@@ -267,7 +295,7 @@ namespace Xharness {
 			
 			args.Add ("--installdev");
 			args.Add (appPath);
-			AddDeviceName (args, CompanionDeviceName ?? DeviceName);
+			AddDeviceName (args, companion_device_name ?? device_name);
 
 			if (mode == "watchos") {
 				args.Add ("--device");
@@ -275,9 +303,9 @@ namespace Xharness {
 			}
 
 			var totalSize = Directory.GetFiles (appPath, "*", SearchOption.AllDirectories).Select ((v) => new FileInfo (v).Length).Sum ();
-			MainLog.WriteLine ($"Installing '{appPath}' to '{CompanionDeviceName ?? DeviceName}'. Size: {totalSize} bytes = {totalSize / 1024.0 / 1024.0:N2} MB");
+			main_log.WriteLine ($"Installing '{appPath}' to '{companion_device_name ?? device_name}'. Size: {totalSize} bytes = {totalSize / 1024.0 / 1024.0:N2} MB");
 
-			return await ProcessManager.ExecuteCommandAsync (Harness.MlaunchPath, args, MainLog, TimeSpan.FromHours (1), cancellation_token: cancellation_token);
+			return await ProcessManager.ExecuteCommandAsync (Harness.MlaunchPath, args, main_log, TimeSpan.FromHours (1), cancellation_token: cancellation_token);
 		}
 
 		public async Task<ProcessExecutionResult> UninstallAsync ()
@@ -298,10 +326,20 @@ namespace Xharness {
 				args.Add ("-v");
 
 			args.Add ("--uninstalldevbundleid");
-			args.Add (BundleIdentifier);
-			AddDeviceName (args, CompanionDeviceName ?? DeviceName);
+			args.Add (bundle_identifier);
+			AddDeviceName (args, companion_device_name ?? device_name);
 
-			return await ProcessManager.ExecuteCommandAsync (Harness.MlaunchPath, args, MainLog, TimeSpan.FromMinutes (1));
+			return await ProcessManager.ExecuteCommandAsync (Harness.MlaunchPath, args, main_log, TimeSpan.FromMinutes (1));
+		}
+
+		bool ensure_clean_simulator_state = true;
+		public bool EnsureCleanSimulatorState {
+			get {
+				return ensure_clean_simulator_state && string.IsNullOrEmpty (Environment.GetEnvironmentVariable ("SKIP_SIMULATOR_SETUP"));
+			}
+			set {
+				ensure_clean_simulator_state = value;
+			}
 		}
 
 		(string resultLine, bool failed, bool crashed) ParseResult (string test_log_path, bool timed_out, out bool crashed)
@@ -334,7 +372,7 @@ namespace Xharness {
 
 					// at this point, we have the test results, but we want to be able to have attachments in vsts, so if the format is
 					// the right one (NUnitV3) add the nodes. ATM only TouchUnit uses V3.
-					var testRunName = $"{AppName} {Variation}";
+					var testRunName = $"{appName} {Variation}";
 					if (xmlType == XmlResultJargon.NUnitV3) {
 						var logFiles = new List<string> ();
 						// add our logs AND the logs of the previous task, which is the build task
@@ -363,24 +401,24 @@ namespace Xharness {
 					return parseResult;
 
 				} catch (Exception e) {
-					MainLog.WriteLine ("Could not parse xml result file: {0}", e);
+					main_log.WriteLine ("Could not parse xml result file: {0}", e);
 					// print file for better debugging
-					MainLog.WriteLine ("File data is:");
-					MainLog.WriteLine (new string ('#', 10));
+					main_log.WriteLine ("File data is:");
+					main_log.WriteLine (new string ('#', 10));
 					using (var stream = new StreamReader (path)) {
 						string line;
 						while ((line = stream.ReadLine ()) != null) {
-							MainLog.WriteLine (line);
+							main_log.WriteLine (line);
 						}
 					}
-					MainLog.WriteLine (new string ('#', 10));
-					MainLog.WriteLine ("End of xml results.");
+					main_log.WriteLine (new string ('#', 10));
+					main_log.WriteLine ("End of xml results.");
 					if (timed_out) {
 						Harness.LogWrench ($"@MonkeyWrench: AddSummary: <b><i>{mode} timed out</i></b><br/>");
 						return parseResult;
 					} else {
 						Harness.LogWrench ($"@MonkeyWrench: AddSummary: <b><i>{mode} crashed</i></b><br/>");
-						MainLog.WriteLine ("Test run crashed");
+						main_log.WriteLine ("Test run crashed");
 						crashed = true;
 						parseResult.crashed = true;
 						return parseResult;
@@ -418,11 +456,11 @@ namespace Xharness {
 				var tests_run = resultLine.Replace ("Tests run: ", "");
 				if (failed) {
 					Harness.LogWrench ("@MonkeyWrench: AddSummary: <b>{0} failed: {1}</b><br/>", mode, tests_run);
-					MainLog.WriteLine ("Test run failed");
+					main_log.WriteLine ("Test run failed");
 					return false;
 				} else {
 					Harness.LogWrench ("@MonkeyWrench: AddSummary: {0} succeeded: {1}<br/>", mode, tests_run);
-					MainLog.WriteLine ("Test run succeeded");
+					main_log.WriteLine ("Test run succeeded");
 					return true;
 				}
 			} else if (timed_out) {
@@ -430,7 +468,7 @@ namespace Xharness {
 				return false;
 			} else {
 				Harness.LogWrench ("@MonkeyWrench: AddSummary: <b><i>{0} crashed</i></b><br/>", mode);
-				MainLog.WriteLine ("Test run crashed");
+				main_log.WriteLine ("Test run crashed");
 				crashed = true;
 				return false;
 			}
@@ -444,7 +482,7 @@ namespace Xharness {
 			CrashReportSnapshot crash_reports;
 			ILog device_system_log = null;
 			ILog listener_log = null;
-			ILog run_log = MainLog;
+			ILog run_log = main_log;
 
 			Initialize ();
 
@@ -453,9 +491,9 @@ namespace Xharness {
 
 			crash_reports = new CrashReportSnapshot () {
 				Device = !isSimulator,
-				DeviceName = DeviceName,
+				DeviceName = device_name,
 				Harness = Harness,
-				Log = MainLog,
+				Log = main_log,
 				Logs = Logs,
 				LogDirectory = LogDirectory,
 			};
@@ -503,8 +541,8 @@ namespace Xharness {
 					ips.Append (ipAddresses [i].ToString ());
 				}
 
-				args.Add ($"-argument=-app-arg:-hostname:{ips}");
-				args.Add ($"-setenv=NUNIT_HOSTNAME={ips}");
+				args.Add ($"-argument=-app-arg:-hostname:{ips.ToString ()}");
+				args.Add ($"-setenv=NUNIT_HOSTNAME={ips.ToString ()}");
 			}
 
 			listener_log = Logs.Create ($"test-{mode}-{Harness.Timestamp}.log", LogType.TestLog.ToString (), timestamp: !useXmlOutput);
@@ -518,7 +556,7 @@ namespace Xharness {
 
 			
 			listener.TestLog = listener_log;
-			listener.Log = MainLog;
+			listener.Log = main_log;
 			listener.AutoExit = true;
 			listener.Address = System.Net.IPAddress.Any;
 			listener.XmlOutput = useXmlOutput;
@@ -533,17 +571,17 @@ namespace Xharness {
 			var timed_out = false;
 
 			listener.ConnectedTask
-				.TimeoutAfter (Harness.LaunchTimeout)
+				.TimeoutAfter (TimeSpan.FromMinutes (Harness.LaunchTimeout))
 				.ContinueWith ((v) => {
 					if (v.IsFaulted) {
-						MainLog.WriteLine ("Test launch failed: {0}", v.Exception);
+						main_log.WriteLine ("Test launch failed: {0}", v.Exception);
 					} else if (v.IsCanceled) {
-						MainLog.WriteLine ("Test launch was cancelled.");
+						main_log.WriteLine ("Test launch was cancelled.");
 					} else if (v.Result) {
-						MainLog.WriteLine ("Test run started");
+						main_log.WriteLine ("Test run started");
 					} else {
 						cancellation_source.Cancel ();
-						MainLog.WriteLine ("Test launch timed out after {0} minute(s).", Harness.LaunchTimeout);
+						main_log.WriteLine ("Test launch timed out after {0} minute(s).", Harness.LaunchTimeout);
 						timed_out = true;
 					}
 				}).DoNotAwait ();
@@ -554,7 +592,7 @@ namespace Xharness {
 			bool? success = null;
 			bool launch_failure = false;
 
-			if (IsExtension) {
+			if (isExtension) {
 				switch (extension) {
 				case Extension.TodayExtension:
 					args.Add (isSimulator ? "--launchsimbundleid" : "--launchdevbundleid");
@@ -592,9 +630,9 @@ namespace Xharness {
 				}
 
 				var systemLogs = new List<CaptureLog> ();
-				foreach (var sim in Simulators) {
+				foreach (var sim in simulators) {
 					// Upload the system log
-					MainLog.WriteLine ("System log for the '{1}' simulator is: {0}", sim.SystemLog, sim.Name);
+					main_log.WriteLine ("System log for the '{1}' simulator is: {0}", sim.SystemLog, sim.Name);
 					bool isCompanion = sim != simulator;
 
 					var log = new CaptureLog (Logs, Path.Combine (LogDirectory, sim.Name + ".log"), sim.SystemLog, entire_file: Harness.Action != HarnessAction.Jenkins)
@@ -607,29 +645,29 @@ namespace Xharness {
 					Harness.LogWrench ("@MonkeyWrench: AddFile: {0}", log.Path);
 				}
 
-				MainLog.WriteLine ("*** Executing {0}/{1} in the simulator ***", AppName, mode);
+				main_log.WriteLine ("*** Executing {0}/{1} in the simulator ***", appName, mode);
 
 				if (EnsureCleanSimulatorState) {
-					foreach (var sim in Simulators)
-						await sim.PrepareSimulatorAsync (MainLog, BundleIdentifier);
+					foreach (var sim in simulators)
+						await sim.PrepareSimulatorAsync (main_log, bundle_identifier);
 				}
 
 				args.Add ($"--device=:v2:udid={simulator.UDID}");
 
 				await crash_reports.StartCaptureAsync ();
 
-				MainLog.WriteLine ("Starting test run");
+				main_log.WriteLine ("Starting test run");
 
 				var result = await ProcessManager.ExecuteCommandAsync (Harness.MlaunchPath, args, run_log, timeout, cancellation_token: cancellation_source.Token);
 				if (result.TimedOut) {
 					timed_out = true;
 					success = false;
-					MainLog.WriteLine ("Test run timed out after {0} minute(s).", timeout);
+					main_log.WriteLine ("Test run timed out after {0} minute(s).", timeout);
 				} else if (result.Succeeded) {
-					MainLog.WriteLine ("Test run completed");
+					main_log.WriteLine ("Test run completed");
 					success = true;
 				} else {
-					MainLog.WriteLine ("Test run failed");
+					main_log.WriteLine ("Test run failed");
 					success = false;
 				}
 
@@ -642,11 +680,11 @@ namespace Xharness {
 							if (line.StartsWith ("Application launched. PID = ", StringComparison.Ordinal)) {
 								var pidstr = line.Substring ("Application launched. PID = ".Length);
 								if (!int.TryParse (pidstr, out pid))
-									MainLog.WriteLine ("Could not parse pid: {0}", pidstr);
+									main_log.WriteLine ("Could not parse pid: {0}", pidstr);
 							} else if (line.Contains ("Xamarin.Hosting: Launched ") && line.Contains (" with pid ")) {
 								var pidstr = line.Substring (line.LastIndexOf (' '));
 								if (!int.TryParse (pidstr, out pid))
-									MainLog.WriteLine ("Could not parse pid: {0}", pidstr);
+									main_log.WriteLine ("Could not parse pid: {0}", pidstr);
 							} else if (line.Contains ("error MT1008")) {
 								launch_failure = true;
 							}
@@ -655,24 +693,24 @@ namespace Xharness {
 					if (pid > 0) {
 						var launchTimedout = cancellation_source.IsCancellationRequested;
 						var timeoutType = launchTimedout ? "Launch" : "Completion";
-						var timeoutValue = launchTimedout ? Harness.LaunchTimeout.TotalSeconds : timeout.TotalSeconds;
-						MainLog.WriteLine ($"{timeoutType} timed out after {timeoutValue} seconds");
-						await ProcessManager.KillTreeAsync (pid, MainLog, true);
+						var timeoutValue = launchTimedout ? Harness.LaunchTimeout : timeout.TotalSeconds;
+						main_log.WriteLine ($"{timeoutType} timed out after {timeoutValue} seconds");
+						await ProcessManager.KillTreeAsync (pid, main_log, true);
 					} else {
-						MainLog.WriteLine ("Could not find pid in mtouch output.");
+						main_log.WriteLine ("Could not find pid in mtouch output.");
 					}
 				}
 
 
 				// cleanup after us
 				if (EnsureCleanSimulatorState)
-					await SimDevice.KillEverythingAsync (MainLog);
+					await SimDevice.KillEverythingAsync (main_log);
 
 				foreach (var log in systemLogs)
 					log.StopCapture ();
 				
 			} else {
-				MainLog.WriteLine ("*** Executing {0}/{1} on device '{2}' ***", AppName, mode, DeviceName);
+				main_log.WriteLine ("*** Executing {0}/{1} on device '{2}' ***", appName, mode, device_name);
 
 				if (mode == "watchos") {
 					args.Add ("--attach-native-debugger"); // this prevents the watch from backgrounding the app.
@@ -682,18 +720,18 @@ namespace Xharness {
 				
 				AddDeviceName (args);
 
-				device_system_log = Logs.Create ($"device-{DeviceName}-{Harness.Timestamp}.log", "Device log");
+				device_system_log = Logs.Create ($"device-{device_name}-{Harness.Timestamp}.log", "Device log");
 				var logdev = new DeviceLogCapturer () {
 					Harness =  Harness,
 					Log = device_system_log,
-					DeviceName = DeviceName,
+					DeviceName = device_name,
 				};
 				logdev.StartCapture ();
 
 				try {
 					await crash_reports.StartCaptureAsync ();
 
-					MainLog.WriteLine ("Starting test run");
+					main_log.WriteLine ("Starting test run");
 
 					bool waitedForExit = true;
 					// We need to check for MT1111 (which means that mlaunch won't wait for the app to exit).
@@ -703,13 +741,13 @@ namespace Xharness {
 						if (line?.Contains ("error MT1007") == true)
 							launch_failure = true;
 					});
-					var runLog = Log.CreateAggregatedLog (callbackLog, MainLog);
+					var runLog = Log.CreateAggregatedLog (callbackLog, main_log);
 					var timeoutWatch = Stopwatch.StartNew ();
 					var result = await ProcessManager.ExecuteCommandAsync (Harness.MlaunchPath, args, runLog, timeout, cancellation_token: cancellation_source.Token);
 
 					if (!waitedForExit && !result.TimedOut) {
 						// mlaunch couldn't wait for exit for some reason. Let's assume the app exits when the test listener completes.
-						MainLog.WriteLine ("Waiting for listener to complete, since mlaunch won't tell.");
+						main_log.WriteLine ("Waiting for listener to complete, since mlaunch won't tell.");
 						if (!await listener.CompletionTask.TimeoutAfter (timeout - timeoutWatch.Elapsed)) {
 							result.TimedOut = true;
 						}
@@ -718,12 +756,12 @@ namespace Xharness {
 					if (result.TimedOut) {
 						timed_out = true;
 						success = false;
-						MainLog.WriteLine ("Test run timed out after {0} minute(s).", timeout.TotalMinutes);
+						main_log.WriteLine ("Test run timed out after {0} minute(s).", timeout.TotalMinutes);
 					} else if (result.Succeeded) {
-						MainLog.WriteLine ("Test run completed");
+						main_log.WriteLine ("Test run completed");
 						success = true;
 					} else {
-						MainLog.WriteLine ("Test run failed");
+						main_log.WriteLine ("Test run failed");
 						success = false;
 					}
 				} finally {
@@ -733,7 +771,7 @@ namespace Xharness {
 
 				// Upload the system log
 				if (File.Exists (device_system_log.FullPath)) {
-					MainLog.WriteLine ("A capture of the device log is: {0}", device_system_log.FullPath);
+					main_log.WriteLine ("A capture of the device log is: {0}", device_system_log.FullPath);
 					Harness.LogWrench ("@MonkeyWrench: AddFile: {0}", device_system_log.FullPath);
 				}
 			}
@@ -748,15 +786,15 @@ namespace Xharness {
 				success = TestsSucceeded (listener_log.FullPath, timed_out, out crashed);
 			} else if (timed_out) {
 				Harness.LogWrench ("@MonkeyWrench: AddSummary: <b><i>{0} never launched</i></b><br/>", mode);
-				MainLog.WriteLine ("Test run never launched");
+				main_log.WriteLine ("Test run never launched");
 				success = false;
 			} else if (launch_failure) {
  				Harness.LogWrench ("@MonkeyWrench: AddSummary: <b><i>{0} failed to launch</i></b><br/>", mode);
- 				MainLog.WriteLine ("Test run failed to launch");
+ 				main_log.WriteLine ("Test run failed to launch");
  				success = false;
 			} else {
 				Harness.LogWrench ("@MonkeyWrench: AddSummary: <b><i>{0} crashed at startup (no log)</i></b><br/>", mode);
-				MainLog.WriteLine ("Test run crashed before it started (no log file produced)");
+				main_log.WriteLine ("Test run crashed before it started (no log file produced)");
 				crashed = true;
 				success = false;
 			}
@@ -784,7 +822,7 @@ namespace Xharness {
 					try {
 						if (pid == 0) {
 							// Find the pid
-							using (var log_reader = MainLog.GetReader ()) {
+							using (var log_reader = main_log.GetReader ()) {
 								string line;
 								while ((line = log_reader.ReadLine ()) != null) {
 									const string str = "was launched with pid '";
@@ -816,7 +854,7 @@ namespace Xharness {
 						if (crash_reason != null) {
 							// if in CI, do write an xml error that will be picked as a failure by VSTS
 							if (Harness.InCI)
-								XmlResultParser.GenerateFailure (Logs, "crash", AppName, Variation, "AppCrash", $"App crashed {crash_reason}.", crash_reports.Log.FullPath, Harness.XmlJargon);
+								XmlResultParser.GenerateFailure (Logs, "crash", appName, Variation, "AppCrash", $"App crashed {crash_reason}.", crash_reports.Log.FullPath, Harness.XmlJargon);
 							break;
 						}
 					} catch (Exception e) {
@@ -830,18 +868,18 @@ namespace Xharness {
 						FailureMessage = $"Killed by the OS ({crash_reason})";
 					}
 					if (Harness.InCI)
-						XmlResultParser.GenerateFailure (Logs, "crash", AppName, Variation, "AppCrash", $"App crashed: {FailureMessage}", crash_reports.Log.FullPath, Harness.XmlJargon);
+						XmlResultParser.GenerateFailure (Logs, "crash", appName, Variation, "AppCrash", $"App crashed: {FailureMessage}", crash_reports.Log.FullPath, Harness.XmlJargon);
 				} else if (launch_failure) {
 					// same as with a crash
 					FailureMessage = $"Launch failure";
 					if (Harness.InCI)
-						XmlResultParser.GenerateFailure (Logs, "launch", AppName, Variation, $"AppLaunch on {DeviceName}", $"{FailureMessage} on {DeviceName}", MainLog.FullPath, XmlResultJargon.NUnitV3);
+						XmlResultParser.GenerateFailure (Logs, "launch", appName, Variation, $"AppLaunch on {device_name}", $"{FailureMessage} on {device_name}", main_log.FullPath, XmlResultJargon.NUnitV3);
 				} else if (!isSimulator && crashed && string.IsNullOrEmpty (crash_reason) && Harness.InCI) {
 					// this happens more that what we would like on devices, the main reason most of the time is that we have had netwoking problems and the
 					// tcp connection could not be stablished. We are going to report it as an error since we have not parsed the logs, evne when the app might have
 					// not crashed. We need to check the main_log to see if we do have an tcp issue or not
 					var isTcp = false;
-					using (var reader = new StreamReader (MainLog.FullPath)) {
+					using (var reader = new StreamReader (main_log.FullPath)) {
 						string line;
 						while ((line = reader.ReadLine ()) != null) {
 							if (line.Contains ("Couldn't establish a TCP connection with any of the hostnames")) {
@@ -851,9 +889,9 @@ namespace Xharness {
 						}
 					}
 					if (isTcp)
-						XmlResultParser.GenerateFailure (Logs, "tcp-connection", AppName, Variation, $"TcpConnection on {DeviceName}", $"Device {DeviceName} could not reach the host over tcp.", MainLog.FullPath, Harness.XmlJargon);
+						XmlResultParser.GenerateFailure (Logs, "tcp-connection", appName, Variation, $"TcpConnection on {device_name}", $"Device {device_name} could not reach the host over tcp.", main_log.FullPath, Harness.XmlJargon);
 				} else if (timed_out && Harness.InCI) {
-					XmlResultParser.GenerateFailure (Logs, "timeout", AppName, Variation, "AppTimeout", $"Test run timed out after {timeout.TotalMinutes} minute(s).", MainLog.FullPath, Harness.XmlJargon);
+					XmlResultParser.GenerateFailure (Logs, "timeout", appName, Variation, "AppTimeout", $"Test run timed out after {timeout.TotalMinutes} minute(s).", main_log.FullPath, Harness.XmlJargon);
 				}
 			}
 
@@ -862,7 +900,7 @@ namespace Xharness {
 
 		public void AddDeviceName (IList<string> args)
 		{
-			AddDeviceName (args, DeviceName);
+			AddDeviceName (args, device_name);
 		}
 
 		public static void AddDeviceName (IList<string> args, string device_name)
@@ -871,6 +909,109 @@ namespace Xharness {
 				args.Add ("--devname");
 				args.Add (device_name);
 			}
+		}
+	}
+
+	// Monitor the output from 'mlaunch --installdev' and cancel the installation if there's no output for 1 minute.
+	class AppInstallMonitorLog : Log {
+		public override string FullPath => copy_to.FullPath;
+
+		ILog copy_to;
+		CancellationTokenSource cancellation_source;
+		
+		public bool CopyingApp;
+		public bool CopyingWatchApp;
+		public TimeSpan AppCopyDuration;
+		public TimeSpan WatchAppCopyDuration;
+		public Stopwatch AppCopyStart = new Stopwatch ();
+		public Stopwatch WatchAppCopyStart = new Stopwatch ();
+		public int AppPercentComplete;
+		public int WatchAppPercentComplete;
+		public long AppBytes;
+		public long WatchAppBytes;
+		public long AppTotalBytes;
+		public long WatchAppTotalBytes;
+
+		public CancellationToken CancellationToken {
+			get {
+				return cancellation_source.Token;
+			}
+		}
+
+		public AppInstallMonitorLog (ILog copy_to)
+				: base (copy_to.Logs, $"Watch transfer log for {copy_to.Description}")
+		{
+			this.copy_to = copy_to;
+			cancellation_source = new CancellationTokenSource ();
+			cancellation_source.Token.Register (() => {
+				copy_to.WriteLine ("App installation cancelled: it timed out after no output for 1 minute.");
+			});
+		}
+
+		public override Encoding Encoding => copy_to.Encoding;
+		public override void Flush ()
+		{
+			copy_to.Flush ();
+		}
+
+		public override StreamReader GetReader ()
+		{
+			return copy_to.GetReader ();
+		}
+
+		protected override void Dispose (bool disposing)
+		{
+			base.Dispose (disposing);
+			copy_to.Dispose ();
+			cancellation_source.Dispose ();
+		}
+
+		void ResetTimer ()
+		{
+			cancellation_source.CancelAfter (TimeSpan.FromMinutes (1));
+		}
+
+		public override void WriteLine (string value)
+		{
+			var v = value.Trim ();
+			if (v.StartsWith ("Installing application bundle", StringComparison.Ordinal)) {
+				if (!CopyingApp) {
+					CopyingApp = true;
+					AppCopyStart.Start ();
+				} else if (!CopyingWatchApp) {
+					CopyingApp = false;
+					CopyingWatchApp = true;
+					AppCopyStart.Stop ();
+					WatchAppCopyStart.Start ();
+				}
+			} else if (v.StartsWith ("PercentComplete: ", StringComparison.Ordinal) && int.TryParse (v.Substring ("PercentComplete: ".Length).Trim (), out var percent)) {
+				if (CopyingApp)
+					AppPercentComplete = percent;
+				else if (CopyingWatchApp)
+					WatchAppPercentComplete = percent;
+			} else if (v.StartsWith ("NumBytes: ", StringComparison.Ordinal) && int.TryParse (v.Substring ("NumBytes: ".Length).Trim (), out var num_bytes)) {
+				if (CopyingApp) {
+					AppBytes = num_bytes;
+					AppCopyDuration = AppCopyStart.Elapsed;
+				} else if (CopyingWatchApp) {
+					WatchAppBytes = num_bytes;
+					WatchAppCopyDuration = WatchAppCopyStart.Elapsed;
+				}
+			} else if (v.StartsWith ("TotalBytes: ", StringComparison.Ordinal) && int.TryParse (v.Substring ("TotalBytes: ".Length).Trim (), out var total_bytes)) {
+				if (CopyingApp)
+					AppTotalBytes = total_bytes;
+				else if (CopyingWatchApp)
+					WatchAppTotalBytes = total_bytes;
+			}
+
+			ResetTimer ();
+
+			copy_to.WriteLine (value);
+		}
+
+		public override void Write (byte [] buffer, int offset, int count)
+		{
+			copy_to.Write (buffer, offset, count);
 		}
 	}
 }
