--- tests/xharness/Collections/IAsyncEnumerable.cs
@@ -1,7 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Xharness.Collections {
-	public interface IAsyncEnumerable {
-		Task ReadyTask { get; }
-	}
-}
