+++ tests/xharness/CrashReportSnapshot.cs
@@ -63,7 +63,7 @@ namespace Xharness
 							var result = await Harness.ProcessManager.ExecuteCommandAsync (Harness.MlaunchPath, sb, Log, TimeSpan.FromMinutes (1));
 							if (result.Succeeded) {
 								Log.WriteLine ("Downloaded crash report {0} to {1}", file, crash_report_target.Path);
-								crash_report_target = await SymbolicateCrashReportAsync (crash_report_target);
+								crash_report_target = await Harness.SymbolicateCrashReportAsync (Logs, Log, crash_report_target);
 								downloaded_crash_reports.Add (crash_report_target);
 							} else {
 								Log.WriteLine ("Could not download crash report {0}", file);
@@ -85,30 +85,6 @@ namespace Xharness
 					}
 				}
 			} while (!crash_report_search_done);
-		}
-
-		async Task<ILogFile> SymbolicateCrashReportAsync (ILogFile report)
-		{
-			var symbolicatecrash = Path.Combine (Harness.XcodeRoot, "Contents/SharedFrameworks/DTDeviceKitBase.framework/Versions/A/Resources/symbolicatecrash");
-			if (!File.Exists (symbolicatecrash))
-				symbolicatecrash = Path.Combine (Harness.XcodeRoot, "Contents/SharedFrameworks/DVTFoundation.framework/Versions/A/Resources/symbolicatecrash");
-
-			if (!File.Exists (symbolicatecrash)) {
-				Log.WriteLine ("Can't symbolicate {0} because the symbolicatecrash script {1} does not exist", report.Path, symbolicatecrash);
-				return report;
-			}
-
-			var name = Path.GetFileName (report.Path);
-			var symbolicated = Logs.Create (Path.ChangeExtension (name, ".symbolicated.log"), $"Symbolicated crash report: {name}", timestamp: false);
-			var environment = new Dictionary<string, string> { { "DEVELOPER_DIR", Path.Combine (Harness.XcodeRoot, "Contents", "Developer") } };
-			var rv = await Harness.ProcessManager.ExecuteCommandAsync (symbolicatecrash, new [] { report.Path }, symbolicated, TimeSpan.FromMinutes (1), environment);
-			if (rv.Succeeded) {;
-				Log.WriteLine ("Symbolicated {0} successfully.", report.Path);
-				return symbolicated;
-			} else {
-				Log.WriteLine ("Failed to symbolicate {0}.", report.Path);
-				return report;
-			}
 		}
 	}
 }
