+++ tests/xharness/Jenkins/TestTasks/NUnitExecuteTask.cs
@@ -76,15 +76,15 @@ namespace Xharness.Jenkins.TestTasks
 			}
 
 			if (is_packageref) {
-				TestExecutable = Path.Combine (DirectoryUtilities.RepositoryRootDirectory, "..", "tools", $"nunit3-console-{nunit_version}");
+				TestExecutable = Path.Combine (Harness.RootDirectory, "..", "tools", $"nunit3-console-{nunit_version}");
 				if (!File.Exists (TestExecutable))
 					throw new FileNotFoundException ($"The helper script to execute the unit tests does not exist: {TestExecutable}");
 				WorkingDirectory = Path.GetDirectoryName (TestProject.Path);
 			} else if (nunit_version [0] == '2') {
-				TestExecutable = Path.Combine (DirectoryUtilities.RepositoryRootDirectory, "..", "packages", "NUnit.Runners." + nunit_version, "tools", "nunit-console.exe");
+				TestExecutable = Path.Combine (Harness.RootDirectory, "..", "packages", "NUnit.Runners." + nunit_version, "tools", "nunit-console.exe");
 				WorkingDirectory = Path.Combine (Path.GetDirectoryName (TestExecutable), "lib");
 			} else {
-				TestExecutable = Path.Combine (DirectoryUtilities.RepositoryRootDirectory, "..", "packages", "NUnit.ConsoleRunner." + nunit_version, "tools", "nunit3-console.exe");
+				TestExecutable = Path.Combine (Harness.RootDirectory, "..", "packages", "NUnit.ConsoleRunner." + nunit_version, "tools", "nunit3-console.exe");
 				WorkingDirectory = Path.GetDirectoryName (TestLibrary);
 			}
 			TestExecutable = Path.GetFullPath (TestExecutable);
@@ -162,7 +162,7 @@ namespace Xharness.Jenkins.TestTasks
 				if (ProduceHtmlReport) {
 					try {
 						var output = Logs.Create ($"Log-{Timestamp}.html", "HTML log");
-						using (var srt = new StringReader (File.ReadAllText (Path.Combine (DirectoryUtilities.RepositoryRootDirectory, "HtmlTransform.xslt")))) {
+						using (var srt = new StringReader (File.ReadAllText (Path.Combine (Harness.RootDirectory, "HtmlTransform.xslt")))) {
 							using (var sri = File.OpenRead (xmlLog)) {
 								using (var xrt = XmlReader.Create (srt)) {
 									using (var xri = XmlReader.Create (sri)) {
