+++ tests/xharness/MonoNativeInfo.cs
@@ -26,9 +26,6 @@
 using System;
 using System.IO;
 using System.Xml;
-
-using Xamarin;
-using Xharness.Hardware;
 using Xharness.Utilities;
 
 namespace Xharness
@@ -157,9 +154,9 @@ namespace Xharness
 
 		public string FlavorSuffix => Flavor == MonoNativeFlavor.Compat ? "-compat" : "-unified";
 		public string ProjectName => "mono-native" + FlavorSuffix;
-		public string ProjectPath => Path.Combine (DirectoryUtilities.RepositoryRootDirectory, "mono-native", TemplateName + FlavorSuffix + ".csproj");
+		public string ProjectPath => Path.Combine (Harness.RootDirectory, "mono-native", TemplateName + FlavorSuffix + ".csproj");
 		string TemplateName => "mono-native" + TemplateSuffix;
-		public string TemplatePath => Path.Combine (DirectoryUtilities.RepositoryRootDirectory, "mono-native", TemplateName + ".csproj.template");
+		public string TemplatePath => Path.Combine (Harness.RootDirectory, "mono-native", TemplateName + ".csproj.template");
 		protected virtual string TemplateSuffix => string.Empty;
 
 		public void Convert ()
