+++ tests/xharness/Targets/TVOSTarget.cs
@@ -1,8 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Xml;
-using Xamarin;
-using Xharness.Hardware;
 using Xharness.Utilities;
 
 namespace Xharness.Targets
@@ -113,7 +111,7 @@ namespace Xharness.Targets
 				MonoNativeHelper.RemoveSymlinkMode (inputProject);
 			}
 
-			var srcDirectory = Path.Combine (DirectoryUtilities.RepositoryRootDirectory, "..", "src");
+			var srcDirectory = Path.Combine (Harness.RootDirectory, "..", "src");
 
 			string project_guid;
 			var mt_nunitlite_project_path = Path.GetFullPath (Path.Combine (srcDirectory, "MonoTouch.NUnitLite.tvos.csproj"));
