+++ tests/xharness/Targets/TodayExtensionTarget.cs
@@ -2,8 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Xml;
-using Xamarin;
-using Xharness.Hardware;
 using Xharness.Utilities;
 
 namespace Xharness.Targets
@@ -63,7 +61,7 @@ namespace Xharness.Targets
 			csproj.SetProjectReferenceInclude ("TodayExtension.csproj", TodayExtensionProjectPath.Replace ('/', '\\'));
 			csproj.FixCompileInclude ("Main.cs", Path.Combine (Harness.TodayContainerTemplate, "Main.cs").Replace ('/', '\\'));
 			csproj.FixInfoPListInclude (suffix, IsGeneratedBclTest ? GeneratedPath : null);
-			TodayContainerGuid = "{" + Helpers.GenerateStableGuid ().ToString ().ToUpper () + "}";
+			TodayContainerGuid = "{" + Harness.NewStableGuid ().ToString ().ToUpper () + "}";
 			ProjectGuid = TodayContainerGuid;
 			csproj.SetProjectGuid (TodayContainerGuid);
 			if (MonoNativeInfo != null) {
