+++ tests/xharness/xharness.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
@@ -21,7 +21,6 @@
     <WarningLevel>4</WarningLevel>
     <ConsolePause>false</ConsolePause>
     <PlatformTarget>x64</PlatformTarget>
-    <LangVersion>8.0</LangVersion>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DebugType>full</DebugType>
@@ -30,7 +29,6 @@
     <WarningLevel>4</WarningLevel>
     <Externalconsole>true</Externalconsole>
     <OutputPath>bin\Release</OutputPath>
-    <LangVersion>8.0</LangVersion>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(RunConfiguration)' == 'Default' ">
     <StartAction>Project</StartAction>
@@ -72,7 +70,6 @@
     <PackageReference Include="Mono.Options" Version="5.3.0.1" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AppInstallMonitorLog.cs" />
     <Compile Include="AppRunner.cs" />
     <Compile Include="BCLTestImporter\BCLTestAssemblyDefinition.cs" />
     <Compile Include="BCLTestImporter\BCLTestImportTargetFactory.cs" />
@@ -82,22 +79,10 @@
     <Compile Include="BCLTestImporter\Platform.cs" />
     <Compile Include="BCLTestImporter\RegisterTypeGenerator.cs" />
     <Compile Include="CrashReportSnapshot.cs" />
-    <Compile Include="Collections\BlockingEnumerableCollection.cs" />
-    <Compile Include="Collections\ILoadAsync.cs" />
-    <Compile Include="Collections\IAsyncEnumerable.cs" />
     <Compile Include="DeviceLogCapturer.cs" />
-    <Compile Include="Execution\MlaunchArguments.cs" />
     <Compile Include="Execution\IProcessManager.cs" />
     <Compile Include="Execution\ProcessManager.cs" />
     <Compile Include="GitHub.cs" />
-    <Compile Include="Hardware\Simulators.cs" />
-    <Compile Include="Hardware\SimDevice.cs" />
-    <Compile Include="Hardware\IDevice.cs" />
-    <Compile Include="Hardware\Device.cs" />
-    <Compile Include="Hardware\Devices.cs" />
-    <Compile Include="Hardware\IHardwareDevice.cs" />
-    <Compile Include="Hardware\ISimulatorDevice.cs" />
-    <Compile Include="Hardware\TCCDatabase.cs" />
     <Compile Include="Harness.cs" />
     <Compile Include="Jenkins\Jenkins.cs" />
     <Compile Include="Jenkins\Resource.cs" />
@@ -147,11 +132,10 @@
     <Compile Include="TestPlatform.cs" />
     <Compile Include="TestProject.cs" />
     <Compile Include="Utilities\Extensions.cs" />
-    <Compile Include="Utilities\Helpers.cs" />
     <Compile Include="Utilities\PlistExtensions.cs" />
     <Compile Include="Utilities\ProjectFileExtensions.cs" />
     <Compile Include="Utilities\StringUtils.cs" />
-    <Compile Include="Utilities\DirectoryUtilities.cs" />
+    <Compile Include="Utilities\TempDirectory.cs" />
     <Compile Include="XmlResultParser.cs" />
   </ItemGroup>
   <ItemGroup>
